fix: Reworked mostFrequentWorld
Reworked mostFrequentWorld method in TaskManager to return the most frequent world of a team, and made sure that if it was tied, that it would be the "special" world, i.e. the nether or the end, depending on which world that/those players are in

Issue: #58
