+++ MinecraftManhunt/src/main/java/manhunt_plus/PluginListener.kt
@@ -1,5 +1,6 @@
 package manhunt_plus
 
+import manhunt_plus.chest_generation.createChest
 import manhunt_plus.chest_generation.generateChestItems
 import manhunt_plus.game_state.AdvancementValue
 import org.bukkit.*
@@ -433,24 +434,6 @@ class PluginListener(var main: PluginMain) : Listener {
 
     }
 
-    /**
-     * Create a chest in the location provided with items
-     *
-     * @param location location where chest should be generated
-     * @param event the block break event
-     */
-    private fun createChest(location: Location, event: BlockBreakEvent, world: World) {
-        val block = world.getBlockAt(location)
-        block.type = Material.CHEST
-        event.isCancelled = true
-        val chest = block.state as Chest
-        val inv = chest.inventory
-        val itemsList = generateChestItems()
-        for (stack in itemsList) {
-            inv.addItem(stack)
-        }
-        Bukkit.broadcastMessage("A chest was generated by " + event.player.name)
-    }
 
     /**
      * Give player a random drop
