+++ MinecraftManhunt/src/main/java/manhunt_plus/chest_generation/ChestUtils.kt
@@ -1,4 +1,10 @@
 package manhunt_plus.chest_generation
+import org.bukkit.Bukkit
+import org.bukkit.Location
+import org.bukkit.Material
+import org.bukkit.World
+import org.bukkit.block.Chest
+import org.bukkit.event.block.BlockBreakEvent
 import org.bukkit.inventory.ItemStack
 import java.util.*
 
@@ -38,4 +44,40 @@ fun generateSupplyDropItems(): MutableList<ItemStack> {
         }
     }
     return itemsToAdd
+}
+
+/**
+ * Creates a chest with items for the supply drop
+ *
+ * @param world the world which the supply drop chest is generated in
+ * @param location the location of the generation
+ */
+fun createSupplyDropChest(world: World, location: Location) {
+    val block = world.getBlockAt(location)
+    block.type = Material.CHEST
+    val chest = block.state as Chest
+    val inv = chest.inventory
+    val itemsList = generateSupplyDropItems()
+    for (stack in itemsList) {
+        inv.addItem(stack)
+    }
+}
+
+/**
+ * Create a chest in the location provided with items
+ *
+ * @param location location where chest should be generated
+ * @param event the block break event
+ */
+fun createChest(location: Location, event: BlockBreakEvent, world: World) {
+    val block = world.getBlockAt(location)
+    block.type = Material.CHEST
+    event.isCancelled = true
+    val chest = block.state as Chest
+    val inv = chest.inventory
+    val itemsList = generateChestItems()
+    for (stack in itemsList) {
+        inv.addItem(stack)
+    }
+    Bukkit.broadcastMessage("A chest was generated by " + event.player.name)
 }
